# src/main/resources/macs-swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "MACS UI rest api endpoints definition.",
    "version": "0.0.7",
    "title": "MACS REST API",
    "contact": {
      "email": "macs_support@aero.bombardier.com"
    }
  },
  "host": "macs-dev.ca.aero.bombardier.net",
  "basePath": "/api",
...
  "schemes": ["https"],
  "paths": {
...
    "/ebom/{projectId}/{sectorId}/{subSectorId}": {
      "get": {
        "tags": ["ebom"],
        "summary": "Fetch ebom data.",
        "description":
          "Fetch ebom table based on specified workspace parameters. The API returned the list of parts that is the merged operation of a) Getting the list of parts from PI and B) Merge the status and quantity (only display remaining qty) with WIP for that PSS (project,sector,subsector).",
        "operationId": "getPartsByProjectSectorSubSector",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of project",
            "required": true,
            "type": "string"
          },
          {
            "name": "sectorId",
            "in": "path",
            "description": "ID of sector",
            "required": true,
            "type": "string"
          },
          {
            "name": "subSectorId",
            "in": "path",
            "description": "ID of subsector",
            "required": true,
            "type": "string"
          },
          {
            "in": "query",
            "name": "partId",
            "description":
              "If set, this limits the search to the provided list of part IDs",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Part"
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "No results"
          }
        }
      },
      "put": {
        "tags": ["ebom"],
        "summary":
          "Update parts status on ebom parts. Only update status. Does not update qty or any other field.",
        "description": "",
        "operationId": "updatePartsByProjectSectorSubSector",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of project",
            "required": true,
            "type": "string"
          },
          {
            "name": "sectorId",
            "in": "path",
            "description": "ID of sector",
            "required": true,
            "type": "string"
          },
          {
            "name": "subSectorId",
            "in": "path",
            "description": "ID of subsector",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Updated quantities and status for provided parts",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Part"
              }
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid parameters"
          },
          "404": {
            "description": "Part(s) not found"
          },
          "405": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/ebom/{projectId}/{sectorId}": {
      "get": {
        "tags": ["ebom"],
        "summary":
          "Fetch ebom table based on specified workspace parameters. NOT USED IN THIS PHASE.",
        "description":
          "Fetch ebom table based on specified workspace parameters.",
        "operationId": "getPartsByProjectSector",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of project",
            "required": true,
            "type": "string"
          },
          {
            "name": "sectorId",
            "in": "path",
            "description": "ID of sector",
            "required": true,
            "type": "string"
          },
          {
            "in": "query",
            "name": "partId",
            "description":
              "If set, this limits the search to the provided list of part IDs",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Part"
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "No results"
          }
        }
      },
      "put": {
        "tags": ["ebom"],
        "summary":
          "Update parts statuses on ebom part. NOT USED IN THIS PHASE.",
        "description": "",
        "operationId": "updatePartsByProjectSector",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of project",
            "required": true,
            "type": "string"
          },
          {
            "name": "sectorId",
            "in": "path",
            "description": "ID of sector",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Updated quantities and status for provided parts",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Part"
              }
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid parameters"
          },
          "404": {
            "description": "Part(s) not found"
          },
          "405": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/ebom/{projectId}": {
      "get": {
        "tags": ["ebom"],
        "summary":
          "Fetch ebom table based on specified workspace parameters. The API returned the list of parts that is the merged operation of a) Getting the list of parts from PI and B) Merge the status and quantity (only display remaining qty) with WIP for that PSS (project,sector,subsector). Only check WIP data based on project ID for all sectors and subsectors.",
        "description":
          "Fetch ebom table based on specified workspace parameters.",
        "operationId": "getPartsByProject",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of project",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Part"
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "No results"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Part": {
      "type": "object",
      "required": ["id", "qty", "status"],
      "properties": {
        "id": {
          "type": "string"
        },
        "number": {
          "type": "string",
          "example": "G25022555-001"
        },
        "qty": {
          "type": "integer",
          "format": "int64"
        },
        "string": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "drawing": {
          "type": "string"
        },
        "revision": {
          "type": "string"
        },
        "authority": {
          "type": "string"
        },
        "parent": {
          "type": "string"
        },
        "type": {
          "type": "string",
          "description": "For example this could be : TOP-ASSSEMBLY"
        },
        "status": {
          "type": "object",
          "description": "status based on which ebom colors are determined",
          "properties": {
            "number": {
              "type": "object",
              "properties": {
                "assignment": {
                  "type": "string",
                  "enum": ["UNASSIGNED", "TOBEASSIGNED", "ASSIGNED"]
                }
              }
            },
            "string": {
              "type": "object",
              "properties": {
                "assignment": {
                  "type": "string",
                  "enum": ["UNASSIGNED", "ASSIGNED"]
                }
              }
            },
            "drawing": {
              "type": "object",
              "properties": {
                "assignment": {
                  "type": "string",
                  "enum": ["UNASSIGNED", "ASSIGNED"]
                }
              }
            },
            "revision": {
              "type": "object",
              "properties": {
                "assignment": {
                  "type": "string",
                  "enum": ["UNASSIGNED", "ASSIGNED"]
                }
              }
            }
          }
        },
        "links": {
          "type": "object",
          "description": "external links",
          "properties": {
            "number": {
              "type": "object",
              "properties": {
                "rightClick": {
                  "type": "string"
                }
              }
            },
            "string": {
              "type": "object",
              "properties": {
                "rightClick": {
                  "type": "string"
                }
              }
            },
            "drawing": {
              "type": "object",
              "properties": {
                "rightClick": {
                  "type": "string"
                }
              }
            },
            "revision": {
              "type": "object",
              "properties": {
                "rightClick": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  }
}